--- frontend/src/components/FileSystem/sample.js
@@ -1,20 +1,0 @@
-[
-  {
-    name: "index.js",
-    filetype: "javascript",
-    content: "const a = Math.random();\nconsole.log(a);",
-    isActive: "false",
-  },
-  {
-    name: "folder",
-    children: [
-      {
-        name: "index.js",
-        filetype: "javascript",
-        content: "const a = Math.random();\nconsole.log(a);",
-        isActive: "false",
-      },
-    ],
-    isActive:"false"
-  },
-];
