+++ frontend/src/main.jsx
@@ -3,18 +3,20 @@ import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.jsx";
 import { BrowserRouter } from "react-router-dom";
-import { Auth0Provider } from '@auth0/auth0-react';
+import { Auth0Provider } from "@auth0/auth0-react";
 
 createRoot(document.getElementById("root")).render(
-  <BrowserRouter>
-    <Auth0Provider
+  <StrictMode>
+  <Auth0Provider
     domain={import.meta.env.VITE_AUTH0_DOMAIN}
     clientId={import.meta.env.VITE_AUTH0_CLIENT_ID}
     authorizationParams={{
-      redirect_uri: "http://localhost:5173/"
+      redirect_uri: "http://localhost:5173/",
     }}
   >
-    <App />
-  </Auth0Provider>,
-  </BrowserRouter>,
+    <BrowserRouter>
+      <App />,
+    </BrowserRouter>
+  </Auth0Provider>
+  </StrictMode>
 );
