+++ package.json
@@ -15,9 +15,12 @@
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
+    "lodash": "^4.17.21",
+    "lodash.debounce": "^4.0.8",
     "mongodb": "^6.14.2",
     "mongoose": "^8.12.1",
     "nodemon": "^3.1.9",
-    "socket.io": "^4.8.1"
+    "socket.io": "^4.8.1",
+    "uuid": "^11.1.0"
   }
 }
